pyproject.toml
[project]
name = "agenticbrowser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain[openai]>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.7",
    "pytest-playwright>=0.7.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
    "langchain-teddynote>=0.5.1",
    "asyncio>=4.0.0",
    "sqlalchemy>=2.0.43",
    "pymysql>=1.1.2",
    "fastapi>=0.118.0",
    "typing>=3.10.0.0",
    "pydantic>=2.11.9",
    "uvicorn[standard]>=0.37.0",
    "aiomysql>=0.2.0",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "pwdlib>=0.2.1",
    "argon2-cffi>=25.1.0",
]
